requirements.txt
flask>=2.3.0
gspread>=5.12.0
google-auth>=2.23.0
rapidfuzz>=3.0.0
sudachipy>=0.6.7
sudachidict-core>=20230927
python-dotenv>=1.0.0
requests>=2.31.0
cachetools>=5.3.0
structlog>=23.2.0
redis>=5.0.0
google-generativeai>=0.3.0
# RAG機能用の依存関係
psycopg2-binary>=2.9.0
pgvector>=0.2.0
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
